Clean up prover a bit more
